+++ e2e-tests/tests/nuxt-3.test.ts
@@ -85,7 +85,15 @@ async function runWizardOnNuxtProject(projectDir: string): Promise<void> {
       'to get a video-like reproduction of errors during a user session?',
     ));
 
-  replayOptionPrompted &&
+  const logOptionPrompted =
+    replayOptionPrompted &&
+    (await wizardInstance.sendStdinAndWaitForOutput(
+      [KEYS.ENTER],
+      // "Do you want to enable Logs", sometimes doesn't work as `Logs` can be printed in bold.
+      'to send your application logs to Sentry?',
+    ));
+
+  logOptionPrompted &&
     (await wizardInstance.sendStdinAndWaitForOutput(
       [KEYS.ENTER],
       'Do you want to create an example page',
@@ -158,6 +166,8 @@ function testNuxtProjectConfigs(projectDir: string) {
       '  replaysOnErrorSampleRate: 1.0,',
       "  // If you don't want to use Session Replay, just remove the line below:",
       '  integrations: [Sentry.replayIntegration()],',
+      '  // Enable logs to be sent to Sentry',
+      '  enableLogs: true,',
       "  // Setting this option to true will print useful information to the console while you're setting up Sentry.",
       '  debug: false,',
       '});',
@@ -172,6 +182,8 @@ function testNuxtProjectConfigs(projectDir: string) {
       '  // We recommend adjusting this value in production, or using tracesSampler',
       '  // for finer control',
       '  tracesSampleRate: 1.0,',
+      '  // Enable logs to be sent to Sentry',
+      '  enableLogs: true,',
       "  // Setting this option to true will print useful information to the console while you're setting up Sentry.",
       '  debug: false,',
       '});',